Cheaper block cache accounting
The runCounter field is used to determine which blocks to retain
during a cache flush. This field counted the number of times an
instruction from the block was executed.

It is simpler and more effective to just count how many time a
block was retrieved from the cache. This reduces the number of
cache flushes and removes an expression from the inner interpreter
loop as well.