Rearrange buffers for faster Glulx memory access.
The original code used two buffers to store Glulx memory:
  gRom: ROM + the initial contents of RAM (potentially memory-mapped)
  gRam: the current contents of RAM only

As a consequence every memory access needed a test to discover
which buffer to use, which is rather unfortunate.

This patch reorganizes the buffers as follows:
  gInitMem: the initial memory contents, same as gRom
  gMem: ROM + the current contents of RAM

Typical ROM size of a Glulx game seems to be between 200 kilobyte
and 2 megabyte, so the extra memory cost is quite reasonable.